Refactor: controller 리팩토링 - 핸들러, 이벤트리스너, 서비스로 코드분리작업 진행
