+++ src/main/java/com/starta/project/domain/liveQuiz/service/LiveQuizService.java
@@ -1,6 +1,7 @@
 package com.starta.project.domain.liveQuiz.service;
 
 
+import com.starta.project.domain.liveQuiz.component.ActiveUsersManager;
 import com.starta.project.domain.liveQuiz.dto.ChatMessageDto;
 import com.starta.project.domain.member.entity.Member;
 import com.starta.project.domain.member.entity.MemberDetail;
@@ -15,14 +16,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.util.HtmlUtils;
 
 import java.time.LocalDateTime;
+import java.util.Set;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class LiveQuizService {
 
-    private final MemberDetailRepository memberDetailRepository;
     private final MemberRepository memberRepository;
+    private final ActiveUsersManager activeUsersManager;
 
     public String findNickName(String username) {
         Member findMember = memberRepository.findByUsername(username).orElseThrow(() -> new IllegalArgumentException("회원 정보를 찾을 수 없습니다."));
@@ -37,7 +39,7 @@ public class LiveQuizService {
         return chatMessage;
     }
 
-    private MemberDetail findMember(String nickName) {
-        return memberDetailRepository.findByNickname(nickName).orElseThrow(() -> new IllegalArgumentException("회원 정보를 찾을 수 없습니다."));
+    public Set<String> getCurrentActiveUsers() {
+        return activeUsersManager.getUniqueNickNames(); // 변경된 부분
     }
 }
